Provide vendor for WP Toolkit software (#22)
Provide a PES vendor to upgrade the WP Toolkit software package.

Due to pecularities with the way the WP Toolkit repository is structured
and operated, a DNF variable is needed in order for the .repo file
provided here to function. That and other miscellaneous fixups are
handled by the leapp repository submitted as
AlmaLinux/leapp-repository#72.