Add preliminary Microsoft vendor
This attempts to allow servers which have Microsoft's YUM repo enabled
to migrate some of the installed software. In particular, Microsoft
actively submits their packages to upstream EL8 repositories, with some
merges and splits.

Since the PES entries need to refer to the upstream repos (in
particular, appstream), and since the PES ID for these change according
to distro, the spec file also has been modified to treat this vendor's PES
file as a template and to insert the proper ID based on the %{dist_name}
being used to build the -data package.